Fix consistency of pixel copy
Previously, we sampled the pixels linearly which could differ in subtle
ways on different machines and graphic stacks. It still can if we intend
to use it for scaling the texture but we will address this at time.
If we copied the same values of pixels in SRGB (and we do not apply any
blending) then the values differ from the ones that result if we copy
them as Unorm. (Uint is not supported for some reason). This is due to
fragments being at (0.5, 0.5) relative to the pixel thus having some
amount of interpolation applied.