Fix staging shader and texture reference for copy size
For a very stupid reason, `shaderc` will silently ignore the entry point
specification in GLSL mode. Thanks, for nothing. Okay, with that
resolved we use a hacky solution for virtually rewriting one of the
methods to main because I want to change shader very little when moving
the naga eventually. Well, probably going to result in a full rewrite to
some other language and this will be in vain but fuck it. shaderc gives
me stupid code, I make more stupid code.

Anyways, we're now using a temporary texture because the validation
layer does not permit us from attaching a texture when we're also
writing to it through its storage binding. That seems sensible (even
though we disable write in pipeline operations) so instead we waste some
memory on a transient texture that the GPU knows we won't read and that
we never initialize, hopefully good enough.

This leaves us with 'only' some weird memory reinterpretation happening
by our round trips but the main code is now GO.

Fly-by fixed a bug where I had resolve the size of a register's texture
by using the DeviceTexture as an index into the ImageBufferMap. Which
doesn't have any idea of device textures and doesn't care so this was
incorrect and finally triggered an out-of-bounds instead of causing more
head-ache in the future. Remove that read in favor of good encapsulation
and we now go through the strongly typed `Register` as an index.