Add test data, remove debugging output
Or better, hide it. I intend to have add some sort of 'debug tap' so
that the same sort of debugging that helped me here can be done in the
eventual version of the library as a help to developers during
optimization or just to gather what is going on in case one is
interested in stats, progress, etc.