Add a 2d distribution constructor
It was all worth it. This remodels the internal (High) paint interface
to more clearly distinguish between purposes. The slightly more general
variants allow using them for painting an initial image that does not
take any inputs. A similar future relaxation will probably take two
inputs for combining them.

In short, it is no easy to define a constructor that works in linear
image color space.