Factor out encoder
This change is for clarity as we will try to formulate any optimization
potential of the encoder as rules of the encoded Low-Op stream only.
This hopefully also makes the immediate program file itself a little
more readable, now focussing on the intermediate representations and
translation into the runnable representation and not the exact mapping.