Rewrite staging with extra layer of grouping
The WebGPU plans to offer storage textures only for textures with 32-bit
texels, not any finer resolution. This makes sense from a compatibility
perspective. However, our 'staging' phase relied on access to textures
at the granularity of single texels. If we are to support small texels
such as 2-bit rgba, 565rgb, 4444yuyv, etc, then we need to map these to
texture formats that aggregate multiple texels to a logical texel for
access.

So, we begin to map 8 bit and 16 bit to 32uint as well. Then another
layer chooses the contained subtexel for decoding or encodes multiple
image texels at the same time.

This is relevant for block mode texels as well where we have a similar
issue but at granularity of pixels and with a potentially lossy
transform instead.