Fix injection mapping wrong color channel
Two parts, a generic decode and one specific to injection.

Since we map the color channel according to its own texel, the
foreground disagreed with the mix argument on which channel to
overwrite. Essentially we had for inject(ColorChannel::R, solid_blue)

background: [...]
foreground: [0, 0, .7, _]
mix:        [1, 0,  0, _]

This would erase the red channel instead of the intended effect.
We fix that by calculating the injected luminance inside the injection
shader through dot product, and setup the product with the source texel
channel in command.

color:      [0, 0, 1, _]

Note: this might provide an optimization where we collapse extract
followed by inject. The former is redundant when measured under the dot
product of the color channel we want to extract.

Now, we also messed up the edianess of color channels during decoding
when texels were aggregated into super-texels.