Move lowering to program
This no longer requires any instantiated device but we carefully only
require the capabilities of the system instead. This also makes it
possible to lower a program directly to an Executable instead of a
started Execution, pushing back providing input buffers.

This path is not yet tested and I don't expect it to (dynamically) work
right now.