Rewrite launching program, for multiple launches
This now allows one to configure an 'Environment' independent from the
Executable. We also now share the instruction list with an `Arc` because
it can be assumed immutable once it leaves the lowering (and later
optimizer) stage. No `Clone` impl for now but this may be necessary to
add if we want to 're-optimize' the program for some reason.