Introduce SrLab2 color spaces
This being so large shows some cracks in the architecture that have
gather some dust in the meantime. It's a bit too hard, and surprisingly
involved, to define these even though they are a mere staging shader
away from other code paths. It's not even possible to convert between
all color spaces equally with this PR..

+276 lines of code, without the shader code