Upgrade wgpu to 0.13
Writes a custom Future based on Arc/Mutex/Condvar to wait on buffer
mapping, which has changed to a callback. Based on description given in
the community chat, wgpu just removed some bad abstraction in favor of
exposing the internal mechanism. Because `.await` was arguably the wrong
mechanic: you'll need to poll the device to ever complete the future and
we want synchronous interfaces to check for progress.

Our new interface is far from perfect. Polling in synchronous way is
still easy to f* up, no guarantee that the callback is invoked often
enough. But the async still works. I'll have to do a second iteration if
we were to reorder the buffer mapping to a prior loop iteration as a
sort of 'prefetching'. No idea how to combine this with RAII yet, as we
can't keep a mutable reference to the buffer then (conflicts with in
step_once's lifetime). But oh well, something for later to figure out.