Prepare dispatching to functions
We'll treat them as frames of instruction, replacing the stack frame of
the current machine state. For this we in particular need a stack in the
first place and a map such that all function refer to only a range of
`Low` instructions instead of all of them. Introduce that to the
execution state in run.rs