Add Knob info for marked buffers
The idea is, we do not want all buffers to be modifiable. That also
comes with an overhead. Such buffers must be mapped with appropriate
usage that will allow an overwrite. Also a buffer must be available for
performing that overwrite. By our current masking rules this buffer can
not be avoided even if we do not use it.

It's still not decided how to translate the update itself. We could
create a fresh buffer with COPY_SRC(mapped_at_creation=true) or we can
try to re-use one buffer across such instructions.