Implement test for image from_buffer constructor
This tests an implementation detail, for now! The error is really really
late if you mess any of this up, i.e. the buffer construction will work
but then the copy will fail. Or the copy works but the pixels are
completely fucked. I'd much rather the parameters for the copy were
explicit and we polyfill more of the alignment / device / backend
specific crap. Ie this should work like blit.